create PipelineExecutor
