+++ modyn/supervisor/internal/grpc/supervisor_grpc_server.py
@@ -2,7 +2,9 @@ import logging
 from typing import Any
 
 from modyn.common.grpc import GenericGRPCServer
-from modyn.supervisor.internal.grpc.generated.supervisor_pb2_grpc import add_SupervisorServicer_to_server  # noqa: E402, E501
+from modyn.supervisor.internal.grpc.generated.supervisor_pb2_grpc import (  # noqa: E402, E501
+    add_SupervisorServicer_to_server,
+)
 from modyn.supervisor.internal.grpc.supervisor_grpc_servicer import SupervisorGRPCServicer
 
 logger = logging.getLogger(__name__)
@@ -11,15 +13,15 @@ logger = logging.getLogger(__name__)
 class SupervisorGRPCServer(GenericGRPCServer):
     @staticmethod
     def callback(modyn_config: dict, server: Any) -> None:
-        add_SupervisorServicer_to_server(
-            SupervisorGRPCServicer(modyn_config), server
-        )
+        add_SupervisorServicer_to_server(SupervisorGRPCServicer(modyn_config), server)
 
     def __init__(self, modyn_config: dict) -> None:
         self.modyn_config = modyn_config
 
         callback_kwargs = {}
-        super().__init__(modyn_config, modyn_config["supervisor"]["port"], SupervisorGRPCServer.callback, callback_kwargs)
+        super().__init__(
+            modyn_config, modyn_config["supervisor"]["port"], SupervisorGRPCServer.callback, callback_kwargs
+        )
 
     def __getstate__(self) -> dict[str, Any]:
         state = self.__dict__.copy()
@@ -37,3 +39,4 @@ class SupervisorGRPCServer(GenericGRPCServer):
             exc_tb (Exception): exception traceback
         """
         super().__exit__(exc_type, exc_val, exc_tb)
+        logger.log("SupervisorGRPCServer exited.")
