+++ modyn/supervisor/internal/pipeline_executor/__init__.py
@@ -1,0 +1,10 @@
+"""Supervisor module. The supervisor initiates a pipeline and coordinates all components.
+
+"""
+import os
+
+from .pipeline_executor import PipelineExecutor  # noqa: F401
+
+files = os.listdir(os.path.dirname(__file__))
+files.remove("__init__.py")
+__all__ = [f[:-3] for f in files if f.endswith(".py")]
